requirements.txt
pymupdf>=1.24.0
ezdxf>=1.3.0
python-frontmatter>=1.1.0
pyyaml>=6.0
click>=8.1.0
rich>=13.0.0
# Stage 2: RAG
chromadb>=1.0.0
sentence-transformers>=3.0.0
# Stage 3: LLM (multi-model)
google-generativeai>=0.8.0
openai>=1.0.0
anthropic>=0.30.0
